refactor: add some changes according to review
