+++ src/components/CartItem/index.tsx
@@ -5,7 +5,7 @@ import {
   removeItemFromCart,
 } from '../../pages/Cart/CartControls/controlFunctions'
 import { CartItem } from '../../types/cartItem'
-import { ProductListProps } from './types/productListProps'
+import { ProductListProps } from './types'
 
 const CartItems = ({ items, setItems }: ProductListProps) => {
   const handleRemoveItem = async (itemId: number) => {
