+++ src/components/Footer/types.ts
