+++ src/components/RangeSlider/types.ts
