+++ src/pages/Contact/index.css
@@ -58,6 +58,20 @@
   cursor: pointer;
 }
 
+.feedback-message {
+  font-size: 14px;
+  margin-top: 10px;
+  text-align: center;
+}
+
+.feedback-message.success {
+  color: green;
+}
+
+.feedback-message.error {
+  color: red;
+}
+
 @media (max-width: 600px) {
   .contact__container {
     width: 90%;
